refactor: Set minimal entry point
Extract control class and types to their own files, adding a lightweight entry point index.ts that only exports everything
