+++ src/components/color/index.ts
@@ -1,9 +1,6 @@
 import gradient from './gradient';
 import list from './list';
-import type { Layer } from 'mapbox-gl';
-import type { ParsedExpression } from '/@/expression';
-
-export type Color = `#${string}`;
+import type { Map, Layer, ParsedExpression } from '../../types';
 
 export default (expression: ParsedExpression<any, any>, layer: Layer) => {
   switch (expression.name) {
