+++ src/expression.ts
@@ -1,5 +1,5 @@
-import type { Expression, ExpressionName } from 'mapbox-gl';
 import { chunk, zip, toPair, toBins } from './utils';
+import type { Expression, ExpressionName, ParsedExpression } from './types';
 
 type Input<T> = T | null | [Input<T>, Input<T>];
 type Stopper = <T>(args: T[]) => [Input<T>, T][];
@@ -14,15 +14,6 @@ const stopper: Record<string, Stopper> = {
 };
 
 const isExpression = (e: any): e is Expression => Array.isArray(e) && !!e.length && typeof e[0] === 'string';
-
-export type ParsedExpression<In, Out> = {
-  name: ExpressionName;
-  stops: [In, Out][];
-  inputs: In[];
-  outputs: Out[];
-  min: number;
-  max: number;
-};
 
 const parse = (input: any): ParsedExpression<any, any> | null => {
   const [name, ...args] = isExpression(input)
